requirements.txt
GitPython
imgaug==0.4.0
jupyter
Keras==2.4.3
keras-vis==0.4.1
nptyping
numpy==1.18.5
opencv-python
openslide-python==1.1.2
pandas==1.1.3
scikit-learn
scipy
Shapely==1.7.1
Sphinx
sphinx_rtd_theme
tensorboard==2.3.0
tensorflow==2.3.1
tqdm
tf-keras-vis==0.5.5

# Image Registration requires histomicstk: $ python -m pip install histomicstk --find-links https://girder.github.io/large_image_wheels

